django_app/requirements.txt
django
python-dotenv
psycopg2
paho-mqtt
channels[daphne]
channels-redis